# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type TaxpayerTaxpayerType =
  (typeof TaxpayerTaxpayerType)[keyof typeof TaxpayerTaxpayerType];

export const TaxpayerTaxpayerType = {
  individual: "individual",
  corporation: "corporation",
  partnership: "partnership",
  estate: "estate",
  trust: "trust",
} as const;

export interface Taxpayer {
  id: number;
  /** Tax Identification Number */
  tin: string;
  registeredName: string;
  tradeName?: string | null;
  taxpayerType: TaxpayerTaxpayerType;
  address: string;
  zipCode: string;
  email?: string | null;
  phone?: string | null;
  /** Revenue District Office code */
  rdoCode: string;
  lineOfBusiness?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaxpayerRequestTaxpayerType =
  (typeof CreateTaxpayerRequestTaxpayerType)[keyof typeof CreateTaxpayerRequestTaxpayerType];

export const CreateTaxpayerRequestTaxpayerType = {
  individual: "individual",
  corporation: "corporation",
  partnership: "partnership",
  estate: "estate",
  trust: "trust",
} as const;

export interface CreateTaxpayerRequest {
  tin: string;
  registeredName: string;
  tradeName?: string | null;
  taxpayerType: CreateTaxpayerRequestTaxpayerType;
  address: string;
  zipCode: string;
  email?: string | null;
  phone?: string | null;
  rdoCode: string;
  lineOfBusiness?: string | null;
}

export type FormSubmissionStatus =
  (typeof FormSubmissionStatus)[keyof typeof FormSubmissionStatus];

export const FormSubmissionStatus = {
  draft: "draft",
  submitted: "submitted",
  filed: "filed",
  amended: "amended",
} as const;

/**
 * Additional form-specific fields stored as JSON
 */
export type FormSubmissionFormData = { [key: string]: unknown } | null;

export interface FormSubmission {
  id: number;
  taxpayerId: number;
  taxpayer?: Taxpayer | null;
  /** BIR form number e.g. 1700, 1701, 1702, 2550M, 2550Q, 1601C */
  formType: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status: FormSubmissionStatus;
  filedDate?: string | null;
  remarks?: string | null;
  /** Additional form-specific fields stored as JSON */
  formData?: FormSubmissionFormData;
  createdAt: string;
  updatedAt: string;
}

export type CreateFormSubmissionRequestStatus =
  (typeof CreateFormSubmissionRequestStatus)[keyof typeof CreateFormSubmissionRequestStatus];

export const CreateFormSubmissionRequestStatus = {
  draft: "draft",
  submitted: "submitted",
  filed: "filed",
  amended: "amended",
} as const;

export type CreateFormSubmissionRequestFormData = {
  [key: string]: unknown;
} | null;

export interface CreateFormSubmissionRequest {
  taxpayerId: number;
  formType: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status: CreateFormSubmissionRequestStatus;
  remarks?: string | null;
  formData?: CreateFormSubmissionRequestFormData;
}

export type UpdateFormSubmissionRequestStatus =
  (typeof UpdateFormSubmissionRequestStatus)[keyof typeof UpdateFormSubmissionRequestStatus];

export const UpdateFormSubmissionRequestStatus = {
  draft: "draft",
  submitted: "submitted",
  filed: "filed",
  amended: "amended",
} as const;

export type UpdateFormSubmissionRequestFormData = {
  [key: string]: unknown;
} | null;

export interface UpdateFormSubmissionRequest {
  formType?: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status?: UpdateFormSubmissionRequestStatus;
  filedDate?: string | null;
  remarks?: string | null;
  formData?: UpdateFormSubmissionRequestFormData;
}

export type FormsSummaryFormsByStatus = { [key: string]: number };

export type FormsSummaryFormsByType = { [key: string]: number };

export interface FormsSummary {
  totalForms: number;
  totalTaxpayers: number;
  formsByStatus: FormsSummaryFormsByStatus;
  formsByType: FormsSummaryFormsByType;
  totalTaxDue: number;
  totalTaxPayable: number;
}

export type ListFormsParams = {
  taxpayerId?: number;
  formType?: string;
  status?: string;
};