# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BIR Form Application API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: forms
    description: BIR form submissions
  - name: taxpayers
    description: Taxpayer management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /taxpayers:
    get:
      operationId: listTaxpayers
      tags: [taxpayers]
      summary: List all taxpayers
      responses:
        "200":
          description: List of taxpayers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Taxpayer"
    post:
      operationId: createTaxpayer
      tags: [taxpayers]
      summary: Create a taxpayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxpayerRequest"
      responses:
        "201":
          description: Taxpayer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"

  /taxpayers/{id}:
    get:
      operationId: getTaxpayer
      tags: [taxpayers]
      summary: Get a taxpayer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Taxpayer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTaxpayer
      tags: [taxpayers]
      summary: Update a taxpayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxpayerRequest"
      responses:
        "200":
          description: Taxpayer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"
    delete:
      operationId: deleteTaxpayer
      tags: [taxpayers]
      summary: Delete a taxpayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Taxpayer deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List all form submissions
      parameters:
        - name: taxpayerId
          in: query
          required: false
          schema:
            type: integer
        - name: formType
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of form submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormSubmission"
    post:
      operationId: createFormSubmission
      tags: [forms]
      summary: Submit a BIR form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormSubmissionRequest"
      responses:
        "201":
          description: Form submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormSubmission"

  /forms/{id}:
    get:
      operationId: getFormSubmission
      tags: [forms]
      summary: Get a form submission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormSubmission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateFormSubmission
      tags: [forms]
      summary: Update a form submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormSubmissionRequest"
      responses:
        "200":
          description: Form updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormSubmission"
    delete:
      operationId: deleteFormSubmission
      tags: [forms]
      summary: Delete a form submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /forms/stats/summary:
    get:
      operationId: getFormsSummary
      tags: [forms]
      summary: Get forms summary statistics
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Taxpayer:
      type: object
      properties:
        id:
          type: integer
        tin:
          type: string
          description: Tax Identification Number
        registeredName:
          type: string
        tradeName:
          type: string
          nullable: true
        taxpayerType:
          type: string
          enum: [individual, corporation, partnership, estate, trust]
        address:
          type: string
        zipCode:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        rdoCode:
          type: string
          description: Revenue District Office code
        lineOfBusiness:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tin
        - registeredName
        - taxpayerType
        - address
        - zipCode
        - rdoCode
        - createdAt
        - updatedAt

    CreateTaxpayerRequest:
      type: object
      properties:
        tin:
          type: string
        registeredName:
          type: string
        tradeName:
          type: string
          nullable: true
        taxpayerType:
          type: string
          enum: [individual, corporation, partnership, estate, trust]
        address:
          type: string
        zipCode:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        rdoCode:
          type: string
        lineOfBusiness:
          type: string
          nullable: true
      required:
        - tin
        - registeredName
        - taxpayerType
        - address
        - zipCode
        - rdoCode

    FormSubmission:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        taxpayer:
          $ref: "#/components/schemas/Taxpayer"
          nullable: true
        formType:
          type: string
          description: "BIR form number e.g. 1700, 1701, 1702, 2550M, 2550Q, 1601C"
        taxableYear:
          type: integer
          nullable: true
        taxablePeriod:
          type: string
          nullable: true
        taxablePeriodEnd:
          type: string
          nullable: true
        grossIncome:
          type: number
          nullable: true
        allowableDeductions:
          type: number
          nullable: true
        taxableIncome:
          type: number
          nullable: true
        taxDue:
          type: number
          nullable: true
        taxWithheld:
          type: number
          nullable: true
        taxPayable:
          type: number
          nullable: true
        penaltiesAndInterest:
          type: number
          nullable: true
        totalAmountDue:
          type: number
          nullable: true
        status:
          type: string
          enum: [draft, submitted, filed, amended]
        filedDate:
          type: string
          format: date-time
          nullable: true
        remarks:
          type: string
          nullable: true
        formData:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional form-specific fields stored as JSON
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - formType
        - status
        - createdAt
        - updatedAt

    CreateFormSubmissionRequest:
      type: object
      properties:
        taxpayerId:
          type: integer
        formType:
          type: string
        taxableYear:
          type: integer
          nullable: true
        taxablePeriod:
          type: string
          nullable: true
        taxablePeriodEnd:
          type: string
          nullable: true
        grossIncome:
          type: number
          nullable: true
        allowableDeductions:
          type: number
          nullable: true
        taxableIncome:
          type: number
          nullable: true
        taxDue:
          type: number
          nullable: true
        taxWithheld:
          type: number
          nullable: true
        taxPayable:
          type: number
          nullable: true
        penaltiesAndInterest:
          type: number
          nullable: true
        totalAmountDue:
          type: number
          nullable: true
        status:
          type: string
          enum: [draft, submitted, filed, amended]
        remarks:
          type: string
          nullable: true
        formData:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - taxpayerId
        - formType
        - status

    UpdateFormSubmissionRequest:
      type: object
      properties:
        formType:
          type: string
        taxableYear:
          type: integer
          nullable: true
        taxablePeriod:
          type: string
          nullable: true
        taxablePeriodEnd:
          type: string
          nullable: true
        grossIncome:
          type: number
          nullable: true
        allowableDeductions:
          type: number
          nullable: true
        taxableIncome:
          type: number
          nullable: true
        taxDue:
          type: number
          nullable: true
        taxWithheld:
          type: number
          nullable: true
        taxPayable:
          type: number
          nullable: true
        penaltiesAndInterest:
          type: number
          nullable: true
        totalAmountDue:
          type: number
          nullable: true
        status:
          type: string
          enum: [draft, submitted, filed, amended]
        filedDate:
          type: string
          format: date-time
          nullable: true
        remarks:
          type: string
          nullable: true
        formData:
          type: object
          additionalProperties: true
          nullable: true

    FormsSummary:
      type: object
      properties:
        totalForms:
          type: integer
        totalTaxpayers:
          type: integer
        formsByStatus:
          type: object
          additionalProperties:
            type: integer
        formsByType:
          type: object
          additionalProperties:
            type: integer
        totalTaxDue:
          type: number
        totalTaxPayable:
          type: number
      required:
        - totalForms
        - totalTaxpayers
        - formsByStatus
        - formsByType
        - totalTaxDue
        - totalTaxPayable