# lib/api-zod/src/generated/types/createFormSubmissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFormSubmissionRequestFormData } from "./createFormSubmissionRequestFormData";
import type { CreateFormSubmissionRequestStatus } from "./createFormSubmissionRequestStatus";

export interface CreateFormSubmissionRequest {
  taxpayerId: number;
  formType: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status: CreateFormSubmissionRequestStatus;
  remarks?: string | null;
  formData?: CreateFormSubmissionRequestFormData;
}