# lib/api-zod/src/generated/types/createTaxpayerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaxpayerRequestTaxpayerType } from "./createTaxpayerRequestTaxpayerType";

export interface CreateTaxpayerRequest {
  tin: string;
  registeredName: string;
  tradeName?: string | null;
  taxpayerType: CreateTaxpayerRequestTaxpayerType;
  address: string;
  zipCode: string;
  email?: string | null;
  phone?: string | null;
  rdoCode: string;
  lineOfBusiness?: string | null;
}