# lib/api-zod/src/generated/types/formSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { FormSubmissionFormData } from "./formSubmissionFormData";
import type { FormSubmissionStatus } from "./formSubmissionStatus";
import type { Taxpayer } from "./taxpayer";

export interface FormSubmission {
  id: number;
  taxpayerId: number;
  taxpayer?: Taxpayer | null;
  /** BIR form number e.g. 1700, 1701, 1702, 2550M, 2550Q, 1601C */
  formType: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status: FormSubmissionStatus;
  filedDate?: Date | null;
  remarks?: string | null;
  /** Additional form-specific fields stored as JSON */
  formData?: FormSubmissionFormData;
  createdAt: Date;
  updatedAt: Date;
}