# lib/api-zod/src/generated/types/formsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { FormsSummaryFormsByStatus } from "./formsSummaryFormsByStatus";
import type { FormsSummaryFormsByType } from "./formsSummaryFormsByType";

export interface FormsSummary {
  totalForms: number;
  totalTaxpayers: number;
  formsByStatus: FormsSummaryFormsByStatus;
  formsByType: FormsSummaryFormsByType;
  totalTaxDue: number;
  totalTaxPayable: number;
}