# lib/api-zod/src/generated/types/taxpayer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { TaxpayerTaxpayerType } from "./taxpayerTaxpayerType";

export interface Taxpayer {
  id: number;
  /** Tax Identification Number */
  tin: string;
  registeredName: string;
  tradeName?: string | null;
  taxpayerType: TaxpayerTaxpayerType;
  address: string;
  zipCode: string;
  email?: string | null;
  phone?: string | null;
  /** Revenue District Office code */
  rdoCode: string;
  lineOfBusiness?: string | null;
  createdAt: Date;
  updatedAt: Date;
}