# lib/api-zod/src/generated/types/updateFormSubmissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIR Form Application API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateFormSubmissionRequestFormData } from "./updateFormSubmissionRequestFormData";
import type { UpdateFormSubmissionRequestStatus } from "./updateFormSubmissionRequestStatus";

export interface UpdateFormSubmissionRequest {
  formType?: string;
  taxableYear?: number | null;
  taxablePeriod?: string | null;
  taxablePeriodEnd?: string | null;
  grossIncome?: number | null;
  allowableDeductions?: number | null;
  taxableIncome?: number | null;
  taxDue?: number | null;
  taxWithheld?: number | null;
  taxPayable?: number | null;
  penaltiesAndInterest?: number | null;
  totalAmountDue?: number | null;
  status?: UpdateFormSubmissionRequestStatus;
  filedDate?: Date | null;
  remarks?: string | null;
  formData?: UpdateFormSubmissionRequestFormData;
}